Fixed block merges. find_category returns nothing if no category.
find_category used to give a KeyError if the supplied data name
was not in the dictionary. Now it returns nothing, so that callers
can gracefully handle data blocks that have a mixture of defined
and undefined data names. Also added guess_category to allow
mixed loops to be assigned a category if possible.

Block merging is still slow but seems robust.